Functionality Fix
- Renamed Home to Index
- Centered Stuff
- Created Components for Expanded List (Compact is pending)
