+++ src/SVG/listing.tsx
@@ -1,0 +1,18 @@
+
+export const CompactList = () => {
+  return (
+    <svg width="24" height="24" viewBox="0 0 30 30" fill="none" xmlns="http://www.w3.org/2000/svg">
+      <path d="M24.875 16.875H5.125C3.25 16.875 2.5 17.675 2.5 19.6625V24.7125C2.5 26.7 3.25 27.5 5.125 27.5H24.875C26.75 27.5 27.5 26.7 27.5 24.7125V19.6625C27.5 17.675 26.75 16.875 24.875 16.875ZM24.875 2.5H5.125C3.25 2.5 2.5 3.3 2.5 5.2875V10.3375C2.5 12.325 3.25 13.125 5.125 13.125H24.875C26.75 13.125 27.5 12.325 27.5 10.3375V5.2875C27.5 3.3 26.75 2.5 24.875 2.5Z" fill="currentColor"/>
+    </svg>
+  );
+};
+
+
+export const ExpandedList = () => {
+  return (
+    <svg width="24" height="24" viewBox="0 0 30 30" fill="none" xmlns="http://www.w3.org/2000/svg">
+      <path d="M27.5238 21.025L23.6113 11.875C22.8988 10.2 21.8363 9.25 20.6238 9.1875C19.4238 9.125 18.2613 9.96249 17.3738 11.5625L14.9988 15.825C14.4988 16.725 13.7863 17.2625 13.0113 17.325C12.2238 17.4 11.4363 16.9875 10.7988 16.175L10.5238 15.825C9.63633 14.7125 8.53633 14.175 7.41133 14.2875C6.28633 14.4 5.32383 15.175 4.68633 16.4375L2.52383 20.75C1.74883 22.3125 1.82383 24.125 2.73633 25.6C3.64883 27.075 5.23633 27.9625 6.97383 27.9625H22.9238C24.5988 27.9625 26.1613 27.125 27.0863 25.725C28.0363 24.325 28.1863 22.5625 27.5238 21.025ZM8.71258 10.4762C9.83312 10.4762 10.9078 10.0311 11.7001 9.23877C12.4924 8.44643 12.9376 7.37178 12.9376 6.25125C12.9376 5.13071 12.4924 4.05606 11.7001 3.26372C10.9078 2.47138 9.83312 2.02625 8.71258 2.02625C7.59204 2.02625 6.51739 2.47138 5.72505 3.26372C4.93271 4.05606 4.48758 5.13071 4.48758 6.25125C4.48758 7.37178 4.93271 8.44643 5.72505 9.23877C6.51739 10.0311 7.59204 10.4762 8.71258 10.4762Z" fill="currentColor"/>
+    </svg>
+  );
+};
+
