+++ src/pages/get-posts/[category].js
@@ -5,19 +5,9 @@ import { moviesListApi } from '../../service/service';
 import ToastMsg from '../../Common/ToastMsg';
 import { Loader } from '../../Common/Loader';
 import { useRouter } from 'next/router';
+import CardExpanded from '../../Common/CardExpanded';
+import { CompactList, ExpandedList } from '../../SVG/listing';
 
-
-function formatBytes(bytes, decimals = 1) {
-  if (!+bytes) return '0 Bytes'
-
-  const k = 1024
-  const dm = decimals < 0 ? 0 : decimals
-  const sizes = ['B', 'KB', 'MB', 'GB', 'TB', 'PB', 'EB', 'ZB', 'YB']
-
-  const i = Math.floor(Math.log(bytes) / Math.log(k))
-
-  return `${parseFloat((bytes / Math.pow(k, i)).toFixed(dm))} ${sizes[i]}`
-}
 
 
 const Latest = () => {
@@ -28,6 +18,7 @@ const categoryId = category ? category.split(':')[1] : "Movies";
   let data = ["Torrents", "Movie", "TV-Show", "Games", "Music", "Anime", "Books", "Other"]
 
   const [movieList, setMovieList] = useState([])
+  const [ListType, setListType] = useState('expanded');
   const [page, setPage] = useState(1)
   const[loader,setLoader]=useState(false)
 
@@ -95,42 +86,19 @@ const categoryId = category ? category.split(':')[1] : "Movies";
       <div>
       </div>
       <br /><br /><br /><br />
-      <div className='w-auto mx-16 px-6 py-8 bg-off-white/10  text-center flex flex-wrap gap-4 place-items-center'>
+      <div className='flex mx-16 justify-between mb-8'>
+        <div></div>
+        <div className="flex bg-off-white/10 rounded-xl">
+          <div className={`px-4 py-2 ${ListType === 'compact' ? 'text-primary bg-primary/30' : ''} rounded-xl`}><CompactList/></div>
+          <div className={`px-4 py-2 ${ListType === 'expanded' ? 'text-primary bg-primary/30' : ''} rounded-xl`}><ExpandedList/></div>
+        </div>
+      </div>
+      <div className='w-auto mx-16 px-6 py-8 bg-off-white/10  text-center flex flex-wrap gap-4 justify-center'>
 
         {movieList?.map((item, index) => {
-          let name = item[`n`];
-          let time = new Date(item[`a`]*1000);
-          console.log(item);
 
           return (
-            <>
-              {<div onClick={()=>{
-                let slug =  name.toLowerCase().trim().replace(/[^\w\s-]/g, '').replace(/[\s_-]+/g, '-').replace(/^-+|-+$/g, '');
-                 router.push(`/post-detail/${item?.pk}/${slug}/`)
-              }} key={index} className="my-1 mx-2 overflow-hidden w-[190px] cursor-pointer py-2 bg-off-white/10 rounded-md flex-col justify-center inline-flex hover:bg-primary/10 border border-off-white/10 hover:border-primary/50">
-                <div className="w-[166px] h-[185px] bg-cover rounded mx-auto justify-center items-center inline-flex" style={{'background-image':`url("${item[`t`] ? item[`t`] : categoryId==="XXX"?"https://i.therarbg.com/xnp.jpg": "https://i.therarbg.com/np.jpg"}")`}}>
-                </div>
-                <br />
-                  <div className="text-off-white text-[12px] h-auto pt-1.5 long-and-truncated font-medium w-fit break-all">
-                  <span>
-                    {name}
-                  </span>
-                </div>
-                <div className="flex text-off-white text-[10px] h-auto pt-1.5 long-and-truncated font-light justify-between">
-                  <span >
-                    {item['c'] || categoryId}
-                  </span>
-                  <span>・</span>
-                  <span >
-                    {time.getDate()}-{time.getMonth()+1}-{time.getFullYear()}
-                  </span>
-                  <span>・</span>
-                  <span >
-                    {formatBytes(item['s'])}
-                  </span>
-                </div>
-              </div>}
-            </>
+            <CardExpanded item={item} categoryId={categoryId}/>
           )
         })}
       </div>
